Double-sided fence lighting without GL_LIGHT_MODEL_TWO_SIDE
Fixes bad performance in Fido level on modern GeForces

+++ Source/3D/3DMath.c
@@ -1908,7 +1908,7 @@ float length, oneOverLength;
 			 (vector3D->y * vector3D->y) +
 			 (vector3D->z * vector3D->z);
 
-	length = sqrt(length);
+	length = sqrtf(length);
 
      //  Check for zero-length vector
 
@@ -1937,7 +1937,7 @@ float length, oneOverLength;
 
 	length = (vector2D->x * vector2D->x) + (vector2D->y * vector2D->y);
 
-	length = sqrt(length);
+	length = sqrtf(length);
 
      //  Check for zero-length vector
